create new group service for group related fetches * change components to use group service * modify group dashboard's timings for fetching data
